Fix intermittent parallel build breakage.
Use an order-only prerequisite to create output directories.

  * This will not trigger a rebuild when the directory mtime changes.

  * This removes the prerequisite from $^, allowing the latter to
    be used in more places.

Add missing prerequisites for the ooxml directory.  One that was missing
broke a parallel build.

Use $^ in a few more places instead of duplicating the prerequisites in
the recipies.

Remove unused prerequisites from the $(wf_GEN_model_QNameToStr_cxx) rule.

Add a prerequisite to the $(wf_GEN_doctok_ResourceIds_hxx) rule that was
in the old makefile.mk.